+++ src/main/scala/com/databricks/labs/automl/pipeline/PearsonFilterTransformer.scala
@@ -1,7 +1,7 @@
 package com.databricks.labs.automl.pipeline
 
 import com.databricks.labs.automl.sanitize.{OutlierFiltering, PearsonFiltering}
-import com.databricks.labs.automl.utils.{AutoMlPipelineUtils, SchemaUtils}
+import com.databricks.labs.automl.utils.{AutoMlPipelineMlFlowUtils, SchemaUtils}
 import org.apache.log4j.{Level, Logger}
 import org.apache.spark.ml.param.{DoubleParam, IntParam, Param, ParamMap}
 import org.apache.spark.ml.util.{DefaultParamsReadable, DefaultParamsWritable, Identifiable}
@@ -26,7 +26,7 @@ class PearsonFilterTransformer(override val uid: String)
 
   def this() = {
     this(Identifiable.randomUID("PearsonFilterTransformer"))
-    setAutomlInternalId(AutoMlPipelineUtils.AUTOML_INTERNAL_ID_COL)
+    setAutomlInternalId(AutoMlPipelineMlFlowUtils.AUTOML_INTERNAL_ID_COL)
     setFieldsRemoved(Array.empty)
     setTransformCalculated(false)
     setDebugEnabled(false)