+++ src/main/scala/com/databricks/labs/automl/pipeline/RoundUpDoubleTransformer.scala
@@ -1,6 +1,6 @@
 package com.databricks.labs.automl.pipeline
 
-import com.databricks.labs.automl.utils.AutoMlPipelineUtils
+import com.databricks.labs.automl.utils.AutoMlPipelineMlFlowUtils
 import org.apache.spark.ml.param.shared.HasInputCols
 import org.apache.spark.ml.param.{Param, ParamMap}
 import org.apache.spark.ml.util.{DefaultParamsReadable, DefaultParamsWritable, Identifiable}
@@ -22,7 +22,7 @@ class RoundUpDoubleTransformer(override val uid: String)
 
   def this() = {
     this(Identifiable.randomUID("RoundUpDoubleTransformer"))
-    setAutomlInternalId(AutoMlPipelineUtils.AUTOML_INTERNAL_ID_COL)
+    setAutomlInternalId(AutoMlPipelineMlFlowUtils.AUTOML_INTERNAL_ID_COL)
     setTransformCalculated(false)
     setDebugEnabled(false)
   }