+++ src/test/scala/com/databricks/labs/automl/pipeline/CardinalityLimitColumnPrunerTransformerTest.scala
@@ -1,6 +1,6 @@
 package com.databricks.labs.automl.pipeline
 
-import com.databricks.labs.automl.utils.AutoMlPipelineUtils
+import com.databricks.labs.automl.utils.AutoMlPipelineMlFlowUtils
 import com.databricks.labs.automl.{AbstractUnitSpec, AutomationUnitTestsUtil, PipelineTestUtils}
 import org.apache.spark.ml.PipelineStage
 import org.apache.spark.sql.DataFrame
@@ -12,7 +12,7 @@ class CardinalityLimitColumnPrunerTransformerTest extends AbstractUnitSpec {
   "CardinalityLimitColumnPrunerTransformerTest" should " should check cardinality" in {
     val testVars = PipelineTestUtils.getTestVars()
     val stages = new ArrayBuffer[PipelineStage]
-    val nonFeatureCols = Array(AutoMlPipelineUtils.AUTOML_INTERNAL_ID_COL, testVars.labelCol)
+    val nonFeatureCols = Array(AutoMlPipelineMlFlowUtils.AUTOML_INTERNAL_ID_COL, testVars.labelCol)
     stages += PipelineTestUtils
       .addZipRegisterTmpTransformerStage(
         testVars.labelCol,